requirements.txt
Django==4.1.3
tencentcloud-sdk-python-cvm==3.0.764
tencentcloud-sdk-python==3.0.764
django-python3-ldap==0.15.3
python-ldap==3.4.3
django-auth-ldap==4.1.0
celery==5.2.7
django-celery-beat==2.4.0
django-celery-results==2.4.0
flower==1.2.0
huaweicloudsdkall==3.1.11
mysqlclient==2.1.1
redis==4.4.0
django-redis==5.2.0
pymysql==1.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
django-cors-headers==3.13.0
chinese_calendar==1.8.0
wxpy==0.3.9.8
pypinyin==0.48.0
wechatpy==1.8.18
kubernetes==26.1.0
pycryptodome==3.17
bs4==0.0.1
qyweixin==0.3.1
paramiko==3.1.0
alibabacloud_alidns20150109==3.0.7
weworkapi==0.1.7
apache-skywalking==1.0.1